refactor(MNT-02): hoist duplicated SafeRefusalAgent into a shared internal type
An identical SafeRefusalAgent stub was copy-pasted (nested) into both BenchOwaspCommand and
BenchMitreCommand. Hoist it to one shared internal AgentEval.Cli.Commands.SafeRefusalAgent and drop
both nested copies. Behavior-preserving (byte-identical logic); verified by build + the 114
OWASP/MITRE tests staying green.

Partial (deferred to vNext): extracting a shared RedTeamCompositeBuilder from OwaspBenchmarkRun and
MitreBenchmarkRun. Those BuildComposite/BuildLeaf/BuildSkippedLeaf/SyntheticEval blocks operate on
different status types (OWASPCategoryStatus vs the MITRE technique status) and are only
near-identical, so a behavior-preserving merge of the security-scoring semantics is a dedicated
refactor best done behind a golden before/after test — there is no behavioral bug today.
